requirements.txt
cython==3.0.11
pandas==2.2.2
pyvista==0.44.1
plotly==5.22.0
scipy==1.13.1
moviepy==1.0.3
cvxopt==1.3.2
nltk==3.8.1
numpy==1.26.4
numpy-stl==3.1.1
rich==13.7.1
loguru==0.7.2
configobj==5.0.8
pytest==8.3.2
pytest-cov==5.0.0
mypy==1.11.1
flake8==7.1.1
tox==4.17.1
tomli==2.0.1
nibabel==5.2.1